refactor search cleanup
