--- lib/character_shower/data/data.dart
@@ -1,3 +1,0 @@
-export '../data/character_models.dart';
-export '../data/result.dart';
-export '../providers/characters_repository.dart';
