--- lib/character_shower/pages/pages.dart
@@ -1,2 +1,0 @@
-export '../pages/character_detail.dart';
-export 'search.dart';
