--- lib/character_shower/widgets/widgets.dart
@@ -1,4 +1,0 @@
-export '../widgets/device_layout.dart';
-export '../widgets/loading_image.dart';
-export '../widgets/search_bar.dart';
-export '../widgets/tablet_layout.dart';
