--- lib/pages/my_home_page.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:show_characters_viewer/flavors.dart';
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(F.title),
-      ),
-      body: Center(
-        child: Text(
-          'Hello ${F.title}',
-        ),
-      ),
-    );
-  }
-}
